Setup register route
Users can now register for the service.
If the user is registered succesfully, json object
with success set to true and the user object
is sent back to the user.
If there is a problem with the registration, an error is thrown
which is handled by the error handling middleware.
If the type of error is either BadRequestError, UnauthorizedError,
or NotFound error, then the error handling middleware sends back
a response with an appropriate status code.
In case any other error is throw a generic error message, with success set
to false is sent back to the user.